+++ crates/cli/src/input.rs
@@ -8,6 +8,8 @@ use std::error::Error;
 use moclib::deser::fits::{MocIdxType, from_fits_ivoa};
 
 
+
+
 #[derive(Debug)]
 pub enum ReducedInputFormat {
   Fits,